add auth handlers (signin/signup) and middlewares

+++ internal/handlers/url/post/delete.go
@@ -8,6 +8,8 @@ import (
 
 	"blog/internal/api/jsonutil"
 	"blog/internal/api/response"
+	"blog/internal/middlewares/auth"
+	"blog/internal/models"
 	"blog/internal/repository"
 	"blog/internal/util/logger"
 	"blog/internal/util/logger/sl"
@@ -19,12 +21,21 @@ type DeleteResponse struct {
 }
 
 type PostDeleter interface {
+	GetPostByID(id int64) (*models.Post, error)
 	DeletePost(id int64) error
 }
 
 func Delete(log logger.Logger, postDeleter PostDeleter) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		log := log.With("fn", "handlers.url.post.Delete")
+
+		authorID, ok := r.Context().Value(auth.UserIDCtxKey).(int64)
+		if !ok {
+			jsonutil.WriteJSON(w, http.StatusInternalServerError,
+				response.Error(http.StatusInternalServerError, "Internal Server Error"))
+			return
+		}
+
 		postID, err := strconv.ParseInt(r.PathValue("id"), 10, 64)
 		if err != nil {
 			log.Info("invalid path value", sl.Error(err))
@@ -33,17 +44,46 @@ func Delete(log logger.Logger, postDeleter PostDeleter) http.HandlerFunc {
 			return
 		}
 
+		post, err := postDeleter.GetPostByID(postID)
+		if err != nil {
+			if errors.Is(err, repository.ErrNotExists) {
+				log.Info("post not found", slog.Int64("post_id", postID), sl.Error(err))
+				jsonutil.WriteJSON(w, http.StatusNotFound,
+					response.Error(http.StatusNotFound, "Post not found"))
+				return
+			}
+			log.Info("post not found",
+				slog.Int64("post_id", postID),
+				sl.Error(err))
+
+			jsonutil.WriteJSON(w, http.StatusInternalServerError,
+				response.Error(http.StatusInternalServerError, "Internal Server Error"))
+			return
+		}
+
+		postAuthorID := post.AuthorID
+		if postAuthorID != authorID {
+			log.Info("forbidden: user is not the author",
+				slog.Int64("post_id", postID),
+				slog.Int64("post_author_id", post.AuthorID),
+				slog.Int64("author_id", authorID))
+
+			jsonutil.WriteJSON(w, http.StatusForbidden,
+				response.Error(http.StatusForbidden, "Forbidden"))
+			return
+		}
+
 		err = postDeleter.DeletePost(postID)
 		if err != nil {
 			if errors.Is(err, repository.ErrNotExists) {
 				log.Info("post is not exists", slog.Int64("post_id", postID))
 				jsonutil.WriteJSON(w, http.StatusNotFound,
-							response.Error(http.StatusNotFound, "Not Found"))
+					response.Error(http.StatusNotFound, "Not Found"))
 				return
 			}
 			log.Error("post delete error", slog.Int64("post_id", postID), sl.Error(err))
 			jsonutil.WriteJSON(w, http.StatusInternalServerError,
-						response.Error(http.StatusInternalServerError, "Internal Server Error"))
+				response.Error(http.StatusInternalServerError, "Internal Server Error"))
 			return
 		}
